Moved system tray commands to help sub menu
